api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
email-validator==2.2.0
